+++ src/main/java/com/nova/narrativa/domain/user/dto/GoogleLoginResult.java
@@ -1,13 +1,14 @@
-package com.nova.narrativa.domain.login.dto;
+package com.nova.narrativa.domain.user.dto;
 
 import lombok.Builder;
 import lombok.Data;
 
 @Builder
 @Data
-public class KakaoLoginResult {
+public class GoogleLoginResult {
 
-    private long id;
+    private Long id;
     private String nickname;
     private String profile_image_url;
+    private String email;
 }
