+++ src/main/java/com/nova/narrativa/domain/user/service/GithubService.java
@@ -1,7 +1,7 @@
-package com.nova.narrativa.domain.login.service;
+package com.nova.narrativa.domain.user.service;
 
-import com.nova.narrativa.domain.login.api.GithubApi;
-import com.nova.narrativa.domain.login.dto.GithubLoginResult;
+import com.nova.narrativa.domain.user.api.GithubApi;
+import com.nova.narrativa.domain.user.dto.GithubLoginResult;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
