+++ src/main/java/com/nova/narrativa/domain/user/service/GoogleService.java
@@ -1,7 +1,7 @@
-package com.nova.narrativa.domain.login.service;
+package com.nova.narrativa.domain.user.service;
 
-import com.nova.narrativa.domain.login.api.GoogleApi;
-import com.nova.narrativa.domain.login.dto.GoogleLoginResult;
+import com.nova.narrativa.domain.user.api.GoogleApi;
+import com.nova.narrativa.domain.user.dto.GoogleLoginResult;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
